requirements.txt
cachetools==4.2.4
certifi==2021.10.8
charset-normalizer==2.0.10
cycler==0.10.0
google-api-core==2.3.2
google-api-python-client==2.34.0
google-auth==2.3.3
google-auth-httplib2==0.1.0
google-auth-oauthlib==0.4.6
googleapis-common-protos==1.54.0
httplib2==0.20.2
idna==3.3
kiwisolver==1.3.1
matplotlib==3.4.2
numpy==1.21.1
oauthlib==3.1.1
Pillow==8.3.1
protobuf==3.19.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.7.0
pyparsing==2.4.7
python-dateutil==2.8.2
requests==2.27.1
requests-oauthlib==1.3.0
rsa==4.8
six==1.16.0
uritemplate==4.1.1
urllib3==1.26.8
